refactor: mv imagegen to vision/imagegen.
