+++ internal/vision/imagegen/dashscope.go
