+++ internal/vision/imagegen/imagegen_test.go
