+++ internal/vision/imagegen/minimax.go
