+++ internal/vision/imagegen/registry.go
