+++ internal/vision/imagegen/vendor.go
